cache JrtFileSystem in ClasspathJrt.jrtFileSystem #2815
to avoid ephemeral memory allocations

https://github.com/eclipse-jdt/eclipse.jdt.core/issues/2815
