+++ org.eclipse.jdt.core/model/org/eclipse/jdt/internal/core/JarEntryFile.java
@@ -53,8 +53,8 @@ public class JarEntryFile  extends JarEntryResource {
 		if (Util.isJrt(root.getPath().toOSString())) {
 			try {
 				Object target = JavaModel.getTarget(root, false);
-				if (target != null && target instanceof File) {
-					return JRTUtil.getContentFromJrt((File) target, getEntryName(), root.getElementName());
+				if (target instanceof File file) {
+					return JRTUtil.getContentFromJrt(file, getEntryName(), root.getElementName());
 				}
 			} catch (IOException e) {
 				throw new JavaModelException(e, IJavaModelStatusConstants.IO_EXCEPTION);
